Synchronise with upstream build script
This also turns off LTO as upstream does not do that by default either
(and neither does the image contain the LLVM gold plugin, though that
could be dealt with easily).

The key consideration here is that by matching upstream behaviour we can
also make sure to produce identical binaries.